feat(day1_part2): try to solve with this
The answer seems not to be correct unless the test file is give me the
right answer

I think there is something wrong with single number in line that is not
explain how to handle or something linked the priority to reversed
numbers, like eightwo is 82 or only 2 because the t of the string is
already used? dont know, need more test anyway!